Add stock price support which necessary scrapers Add Selenium in docker-compose.yml to scrap data from Gpw and NewConnect Fix wallet query for averagePurchasePrice in AssetRepository Change update asset method from PUT to PATCH Add currency field in FinanceDetail and change necessary classes and tests Remove old fetch method which /currency endpoint Change edit-asset component to support patch instead of post

+++ backend/services/wallet/src/main/java/pl/muybien/asset/dto/AssetGroupDTO.java
@@ -12,7 +12,7 @@ public record AssetGroupDTO(
         String uri,
         AssetType assetType,
         BigDecimal count,
-        Double averagePurchasePrice,
+        BigDecimal averagePurchasePrice,
         String currency,
         String customerId
 ) {
